payment-service/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8